refactor: 유틸파일 관심사 분리
makeEpicWithDate 라는 함수는 date와 관심사가 다른 함수이므로 epic.ts 에 분리함
