+++ client/src/lib/utils/date.ts
@@ -1,14 +1,3 @@
-import { EpicType, EpicWithString } from '@/types/epic';
-
-export const makeEpicWithDate = (epicWithString: EpicWithString): EpicType => {
-  return {
-    id: epicWithString.id,
-    name: epicWithString.name,
-    startAt: new Date(epicWithString.startAt),
-    endAt: new Date(epicWithString.endAt),
-  };
-};
-
 export const getYMD = (date: Date) => {
   const year = date.getFullYear();
   const month = date.getMonth();
